[feature1283Strafing] + FiringDisplay and other UI elements for strafing
FiringDisplay now has a Strafe button (also added the menu item as well).  When
Strafe is clicked, the status label channges with information on strafing.  The
click behavior also changes: clicking will select a target Coord for the
strafing run and if the coord is valid it will be highlighted in yellow. The 
UnitDisplay will also display to-hit information for all Entity targets that
will be hit by the strafing run.  

Also updtaed l10n entries for strafing.