+ Bug: MovePath.findPathTo can return a path where MoveStep.parent isn't properly updated in each step
MoveStep.parent is used in some instances to determine state for a MoveStep,
such as determining if the MoveStep is in an end position.  When a MovePath
gets extended, the parent state of each MoveStep isn't updated, and this can
cause issues with the path that gets returned.  I added code just before the 
path is set to iterate through each step and set the parent.