+ Bug: KeyBinds non-functional and all set to -1 if defaultKeyBind.xml not found
The current defaultKeyBinds.xml is now encoded in KeyCommandBind.java, so all
keys have a default in case no file is found.  I also updated the XML output
code so that it writes the symbolic name for each key bind in comments (as the
keycode are generally not human readable).