+ Bug: Fixed issue with BoardView1.redrawMovingEntity
It looks like BoardView1.redrawMovingEntity intends to remove the current
EntitySprite and IsometricSprite for the moving unit, however it fails to
do this.  It looks like the intent is to draw a transparent version of the
Entity at the end of the movement, and move an opaque version of the Entity.
However, the transparent version is overwritten by the Entity/Isometric Sprite
because they aren't properly removed.  This has been fixed.