+ Separated ammo dumping and mode switching
I found that the if statement in SystemPanel.valueChanged for the when the slot
value is changed links the ammo dumping button and the mode selection drop down
This essentially means that ammo can either be dumped, or have it's mode
changed.  If dumping is allowed, mode changing isn't.  I can't see a reason for
this, so I separated the if/else into two, one for dumping and one for mode
selection.

Additionally, ammo dumping is only allowed for ammo without instant mode
switching.  Currently there aren't yet any AmmoTypes that allow instant mode
switching, but I can't see a reason for this requirement so I removed it.

Finally, I made it so that equipment that is being dumped cannot have its mode
changed.