+ Bug: Dug-in field gunners would fail to fire if secondary arc changed
Infantry.isSecondaryArcWeapon would allow the field gunners to torso twist,
so the Client allows it, but then when Server processes the TorsoTwistAction
it calls Infantry.canChangeSecondaryFacing which had an additional check that
the infantry aren't dug in.  I can't find anything in the rules that indicate
digging in has any effect on field gunners changing their weapon facing, so
I removed this check.