+ Encoded terrain feature elevation into Terrains
There's now a static method in Terrains that will return the number of
elevations a given terrain type at a specific terrain level rises above the
surface of a hex (ie, 2 elevations for light/heavy woods).  This then is called
by a new method in Terrain for returning the values for a specific instance of
a Terrain.  I then added Hex/IHex.maxTerrainFeatureElevation to return the max
terrain feature elevation for a given Hex.  This allowed me to rewrite
Hex.ceiling down to pretty much one line.  This satisfies the TODO in Hex.