+ Bug: Hammerhead, roll, half roll, maneuvers allowed free facing changes
If the isManeuver state of MoveStep is set, then Aeros can make turns.  This
state was set for Hammerhead, roll, and half roll although these maneuvers
shouldn't allow facing changes.  The free facing changes for maneuvers must
stay in order for Immelmann's and Split-S's to work correctly.  Currently,
MoveStep.isManeuver() is only used to check for free turns.  So, to fix the
issue, I just made it so that Hammerheard, roll, and half-roll don't set the
isManeuver state.