+ Bug [#4219]: Added call to MovePath.compile for dest path in pathplanner
There's some state for the move path that is only set for the last valid step
in a MovePath, for instance the danger flag when landing a jump in heavy woods
with the optional rule on.  The path planner ends up expanding the path past
this point, which changes the danger state.  In order to ensure that the
returned path has the proper end step state, it needs to be compiled before
being returned.  This ends up taking extra time, but helps make sure that the
path is correct.

I also added a flag to MovePath.compile that determines whether the path should
be clipped to possible steps.  This needs to be done when the path is compiled
for purposes of making a move, but shouldn't be done when planning a path.