+ Bug [#4219]: Removed reusing of StepSprite
Now that we're using shallow copies of MovePath, MoveStep.canReuse is
meaningless.  If we're considering reusing a step, then the old step is a
shared reference with the new step, so this method always returns true.

This is an unfortunate side effect, but I'm guessing that the speed-ups
gained in the path planner far outweight the speed-ups gained by reusing
sprites.  Plus, with the other rendering speed-ups of the past year or so,
I don't think this is as important as it once was.