+ Bug [#4219]: Re-verted some changes to MoveStep.getMovementType
Previously, MoveStep.getMovementType had some code to check whether the step
was at the end of a MoveStep, and if it is, it would check MoveStep.isLegalEndPos
to determine if it should return MOVE_ILLEGAL or return MoveStep.moveType. Turns
out this check was really important.

The code is setup to not set MoveStep.moveType to MOVE_ILLEGAL for certain moves
that are conditionally illegel, such as stacking violations (you can move
through an ally, but not stop on them).  If moveType was set to MOVE_ILLEGAL
in these cases, it gets much hairier than if the step is conditionally illegal
only when it is the end position.

So, I added back this check, and since the MoveStep itself cannot know if it's
an end step, I added a boolean isEnd flag to MoveStep.getMovementType.

I then updated all the references to this to send in the proper value of the
flag.  In many of the cases, the last step in the path is used, so it was
simple.  In other cases, I added MoveStep.isEndPos, which determines if a
supplied MoveStep is an End position.

It's important to not that being an end position is more than just being the
last step.  The end position must be the last legal step in a path, otherwise
it's possible to add additional illegal steps and make a formerly illegal
step legal.