+ Bug: Accidental fall from above check fails when target is standing in water
Game.getAffaTarget was using entity.getElevation() == 0 as a check to ensure
that the target entity was on the surface of a hex.  However, if the target is
standing in depth 1 water, their elevation would be -1, and not 0, and they
should still be a valid affa target.  I updated it so elevation == 0 ||
elevation == -hex.depth.  So, standing on the surface or standing on the floor.
(Hex.floor is an absolute level and not a relative elevation, so is not
applicable here).