Adjusted error-handling in CamoChoiceDialog
When the camo cannot be loaded, a null ImageIcon is returned.  Previously, a
stack trace would be printed to the logfile.  I replaced this with a more
useful message.  This helps make it clear what's happening, and also makes it
more clear that it's a handled exception and not an unhandled one.
Also, a JDialog was displayed, however this could be displayed outside of the
ChatLounge and hence be confusing to players, so I made it so the dialog only
displays if the CamoChoiceDialog is visible.