+ Bug: /load command requires players to disconnect/reconnect after game loads
Server ensures that the connection ID matches the player ID.  If the /load
command is used, then a list of players may already be connected (with assigned
connection IDs).  When the game is loaded, it loads a new list of players, and
so the match between player ID and connection ID is not guaranteed to hold.

I added to code that will correct this.