+ Bug: Issues with updating double-blind visuals information
Sometimes, when a unit moved, the visual indicators for all units wouldn't
be updated properly.  This ultimately came down to Server.updateVisibilityIndicator.
This method updates every unit with what players have units that can see the unit.
The problem was the logic that determined if clients needed to be updated.
This logic was working on boolean flags, which would fail when you got to games
with more than 2 players.  I decided that sending the visibility indicators is
small overhead and so it should always be sent, unless an entityUpdate is required.
An entityUpdate is fired if there is some client that now sees/detects the entity
that wasn't able to see/detect it before.