+ Fixed a crash with APDS implementation
Would crash with an NPE when any missile attack was fired and not in range
of an APDS.  I used a "for each" construction over a List<Mounted>.  I
had thought that if the list was null, it wouldn't iterate over it.
Turns out, it NPEs... So, I added a check to ensure that a list exists
before trying to iterate over it.