+ Added KeyBind for "Done" action
Previously, Ctrl-Enter could be used to press the Done button, however this was
using the swing InputMap and ActionMap.  The 'when in focused window' input map
was used, and this would lead to cases where the key press would get ignored.

By using the KeyBind system, not only can this key be remapped, but it should
work regardless of what window is in focus.

To do this in the most compact way, I made it so the key mapping is registered
in AbstractPhaseDisplay, which required changing the constructor.  This
probably should have been done in the past, as it removes a bit of repeated
code.