Consistenty between Game.addEntity and Entity.setBAGrabBars
When Game.addEntity is called, it calls setBAGrabBars for Mech and Tank, which
will add BattleArmorHandles if the option is on, or ClampMounts if the option
isn't on.  After this happens, there's code that will add ClampMounts to Mechs
and Tanks (but not VTOLs), if they aren't omnis and don't already have handles
(which they always should because of the previous code).  I removed the !VTOLs
check, to make it consistent with Entity.setBAGrabBars.