+ Bug: Answering 'No' to Vertical Takeoff PSR Nag doesn't clear move path
I made it so that after ready() is called with the added VTAKEOFF step
(MovementDisplay line 4232), clear() is called.  If ready() doesn't send
the movement for any reasons (ie, saying No to the PSR nag), then the
move path is cleared.  Any prior move path was already cleared before the
VTAKEOFF step is added, so we won't lose anything.  Previously, the user would
have had to clear the movepath manually, and there was no visual indicator this
was the case.  Now the path is cleared for them.