+ Bug: Missing 'in crater' penalty for vertical take-off (TW pg 88)
The TODO indicated that TW has no definition of what a crater is.  There's no
formal definition, but when a DS vertically lands off of pavement, it reduces
level of all hexes it lands on by 1.  This would create a "crater".  Based on
this, I think it's reasonable to define a crater as 'all adjacent hexes to the
landed dropship have a higher level than the hexes the dropship is landed on.'
I have implemented this check.