+ Code refactoring in BoardView1
Sprite.getBounds() can potentially cause some computation to happen: in Sprite,
it just returns the bounds directly, however, in many of the subclasses, some
computation is performed and potential new objects can be created.  Because of
this, this method should only be called when necessary.  I refactored some code
to called and store the value, rather than making 3 separate calls.