+ Feature: Unit names in reports are now hyperlinks, clicking them displays unit
I thought that it would be nice to be able to view units during the report
phases, so I made it so Report.addDesc(Entity) makes the entity name a
hyperlink.  I then updated ReportDisplay to implement HyperLinkListener and
when a hyper link is clicked, it looks for #entity:<entityId> and will use
that to display the clicked upon Entity.