+ Setup framework for making pointblank shots
The Server.processMovement is now aware of hidden units and will check for
stacking violations with hidden units, and moving adjacent to a hidden unit.
If adjacent movement happens, a pointblank shot client feedback request will
be made, and the client controlling the hidden unit will be asked if it wants
to make a PBS.  The framework for sending a collection of AttackActions is
there, however for now nothing happens if 'Ok' is clicked on the dialog, right
now the PBS can only be declined.