+ Added PointblankShotDisplay, and set it up to display during PBS resolution
I don't like this solution, as I feel like it gives the player the indication that they can
take a long time making the PBS decision, when in reality they shouldn't.  The server
is basically halted waiting for a response, and so no chat messages are relayed, and
saving the game at this point would be detrimental.  The PBS CFR is very much a temporary
state that should end quickly.  With this in mind, I'm going to favor a different UI choice.