+ Bug: Server appears unresponsive during a client feedback request
The server blocks on the cfrPacketQueue while waiting for a feedback request,
and any non-CFR packets are just queued for processing after the CFR is
complete.  This means that clients connecting/disconnecting and players
sending chat messages will not be processed until after the CFR is complete.
To get around this, I've partitioned the packets into three categories:
CFR packets, which are enqueued into the cfrPacketQueue, regular packets
which are enqueued into the packetQueue, and packets that should be processed
immediately by the connection listener.  These packets include things related
to connecting/disconnecting and chat.