+ Added multiple font colors, button default/disabled/activated colors can now be specified
SkinSpecifications now have a list of colors, instead of one.  Updated the code
to read/write this list of colors.  MegaMekButtons will now use the first entry
as the default button text color, the second entry as the disabled text color
and the third color will be the activated color.