+ Added UnitDisplaySkinSpecification
UnitDisplay can now have the images it uses defined in
UnitDisplaySkinSpecification, which can be accesed from SkinXMLHandler.
The previous values for the images are encapsulated as defaults within
UnitDisplaySkinSpecification, so if it's not defined in a skin spec, there
won't be any issues.

Right now, SkinXMLHandler can't read or write the values to XML, but that will
come soon.