+ Added skinning for UnitDisplay
I had to update a all of the unitDisplay panels as well as all of the MapSets
to use the UnitDisplaySkinSpecification values instead of hard-coded values.

I then added UnitDisplaySpecPanel, which is a variant of SkinSpecPanel.
SkinSpecEditor was updated to display either SkinSpecPanel or
UnitDisplaySpecPanel based upon what component is selected.

SkinEditorMainGUI now display the UnitDisplay, loads a test entity to display.

Also added support for UnitDisplaySkinSpecification in SkinXMLHandler.