+ Fixed in EntitySprite.updateLabel
Game.getEntitiesVector(1) was being used, when the (2) version should be used.
The second parameter indicates whether the targetability of the unit should be
ignored.  The default is false.  This would cause an issue when deploying: if
a unit is deployed in the hex of the labels, the units position wasn't
considered because it wasn't targetable.  By using the second parameter with
a value of  true, label position is properly updated.