+ Adjusted polygons for UnitDisplay tab strip
In a previous commit, I had attempted to allow custom tab strip images, however
I failed to realize that there's a poly specification in addition to the imgs.
I found that the current hard coded polygons seemed to be more detailed than
they needed to be (there were points I could remove that appeared to have
no visual effect).

Now, these polgyons are defined based on the image sizes, so it should be
possible to use different sized images, and it should be possible to have
varying widths as well.