+ Fixed an issue with stale images
I was finding that, if I changed an image on disk without changing the path,
when it comes to the UnitDisplay tab strip, those images wouldn't update.
Looks like this comes down to Toolkit.getImage(...) which will attempt to
use the same image for multiple calls to the same location.  I needed to
add flush calls to the retrieved images to ensure that they were up-to-date.