+ Bug: Floor deploy dialog for Infantry in buildings could select wrong floor
The elevation used was based on the index of the floor name selected.  However,
if stacking violations would remove certain floors, these indices were no
longer valid for elevation.  I added a second ArrayList to keep track of the
floor elevations as well as floor names.

Also, I added a JOptionPane for when there are no valid floors in the hex,
previously the user would click on the hex and nothign would happen. Now they
will get a message saying the unit cannot deploy in that hex.