Replacing Coords.hashCode() with something suitable for a PRNG
This repurposes the HashCodeUtil class to have a central place for specialised hashes. To do so cleanly, the following trivial changes had to be made:

* Server.EntityTargetPair's hashCode() and equals() changed to use the standard Objects.hash() and Objects.equals() methods
* The Targetable interface made to require both methods, since it's used in the class above

Also, the TestCoords class (which seems otherwise unused) was cleaned up and the "get coordinates from hash code" method in Coords removed, since it didn't make much sense to begin with.