+ Bug: Entity tooltip could have incorrect movement mod with TO standing still
The i18n entry for no movement was presuming a TMM of 0.  This isn't the case
with the TO standing still optional rule on.  I just changed the i18n entry to
take a parameter for the movement modifier, and set the value to what is
returned by Compute.getTargetMovementModifier.