Fixed focus issue with path text fields
Previously, typing in an updated file name would cause the skin editor panel to
lose  focus.  There were two things I had to do to fix this.

First, I updated SkinSpecEditor.notifySkinChanges to take a flag that
determines if the skinEditPanel should be updated.  This allows me to notify
skin changes, but not recreate the whole skinEditPanel if it's not necessary
(like for file path changes).

Secondly, I had to add some hacky code to SkinEditorMainGUI.updateBorder. The
main problem is that there's no way to update hte skin for the UnitDisplay
without recreating it, but if I just recreate the UnitDisplay without
recreating the JDialog that houses it, the UnitDisplay doesn't work right.
In the end, I came up with a hacky way to force it to work.