+ Issue with individual camo and "No Camo"
The CamoChoiceDialog has a "No Camo" category, which is used for picking a camo
that is a color.  What this does is change the Player object's color index.
This only makes sense for picking a Player's camo and not an Entity's.
When picking individual camo, it will show all colors, but selecting any of
them will cause the unit to display the players colors (basically, there's no
mechanism supporting colors different from the players color).

To make this more clear, I made it so the CamoChoiceDialog will only display
the Players color when setting individual camo.