+ Bug: MiniMap selected unit highlighting persists after turn ends
This is because the MiniMap highlighting is based upon
ClientGUI.selectedEntityNum, which gets set in the various
StatusBarPhaseDisplay implementations (aka FiringDisplay, MovementDisplay),
in the beingMyTurn() method.  However, it is never cleared.  I added code
in the endMyTurn() method to clear this value.  It appears to me that it was
intended to be cleared but never was because it had no effect, until now.