Initial rendering code for board backgrounds
If a board has a background image, and a hex uses the transparent theme,
instead of drawing the hex normally, the image is pulled from the background
image.  base, super, and ortho aren't rendered.

There are still some glitches with the rendering.  It looks like the whole hex
square is rendered, meaning the hex border lines get overwritten, and the
images aren't scaling properly.  Next steps are to work on these bugs.