Fixed rendering issues
First, there was the issue with the background chunks being square instead of
hex shaped.  This was because I used Component.createImage, and instead needed
to use a BufferedImage.  This is because AlphaCompositing is used to limit the
hex image to just the hex, by first drawing the hex mask, and then using
SRC_ATOP alpha compositing when the hex image is drawn.  Componenet.createImage
images don't support alpha channels/transparency, so  this method fails.

Next, the scaling issue... What was happening was scaling was being done twice.
I was scaling the hex image when retrieving the "transparent base", and then
that was getting scaled again.  So, I removed the scaling code from
getTransparentImage (which probably should have its function name changed).