Fixed scaling issue again
I do need to scale the board background, like I was doing originally, but I
also need to not scale the output again.  Now things are scaling nicely,
however I want to cache all of the scaled background images when the board
is loaded, to save time in rendering later.