Removed unnecessary references to "visibility" option in GameOptionsDialog
Looks like there used to be a "visibility" option that was used to set how far
units could see in double-blind.  This option was removed in c4e850 when a
slew of TO options were added, however the references to this option were not
removed from GameOptionsDialog.  This leads to an NPE when the "Double Blind"
option is clicked.