+ Issue #243: Removed superfluous high gravity jump PSR
If 'mechs jump in high gravity, they must make a PSR and if they fail they
take 1 pt of damage for each walk MP they have lost due to gravity. This roll
was often being made superflously, for units that would take 0 damage on a
failed roll.  I added a condition to the roll check to ensure that the roll
is only made if damage is actually dealt.