+ Feature: Added "Swarmed" entry to unit tooltips
I noticed that the only way to tell if a unit was swarmed was from the
"SWARMED" status on the icon.  Right now, if multiple stacked units have
statuses, they can get overlapped.  I added an entry to the tool tip for
an entity, so it's easier to tell what is swarmed, and by what unit.