Added TestEntity.setPrecision
This allows a double value to be rounded to an arbitrary precision (# of
decimal places).  When computing weights for validation, we often run into
floating point representation problems, so we should generally be careful to
round values to a specific precision to present problems.

An example in this case was the Galaport.  Computing engine weight, for SVs
we take the engine base value, movement factor, engine weight mult and
entity weight and multiple them together.

This yields 0.025 * 8 * 1.5 * 200.  0.025 * 8 * 1.5 = 0.3 which in floating
point is .30000000000000004.  This leads to a calculation of 60.000000000000001
which then gets rounded up to the nearest half-ton, for 60.5, while the rules
expect 60 tons.  In this case, by setting the precision to 4 decimal points,
we can get 60.0000.