+ Issue #284: Prevented null images for base, super, ortho
I added checks for images returned by HexEntry.getImage to see if the results
are null, and if they are a new blank image is created.

Combined with the logging of loading hex tilesets, and HexSetTest, this should
eliminate issues.  It's far easier to be sure that our tilesets are valid,
and if there is an issue it's logged.  The last step is to make the code more
robust to issues as they crop up.

Now, instead of receiving an NPE, nothing will be drawn.  This will look odd
to the user, but they can continue on, and if they look in the log the problem
can be figured out.