+ Bug: Manually shutdown units skipped during movement with option turned on
If a unit is manually shutdown and the skip ineligible during movement option
is on.  The problem is that, the conditional for returning false included
!Entity.isActive() || ... Entity.isActive will always return false if the unit
is shutdown or manually shutdown.  Because of this, the eligible for movement
check always returned false when a unit was manually shutdown.  The logic for
isActive needs to be adjusted for this eligibility check.  Instead of calling
Entity.isActive, I pulled out the logic for that method and adjusted it (since
it's pretty short).