Bug: Hotkeys not ignored in some cases (ie: scrolling when save menu is up)
Added BoardEditor.ignoreHotkeys, similar to how ClientGUI works.  I also made
it so that when any of the JTextFields have focus, the hotkeys are ignored.
This is most convenient for themes, however the problem is the user has to lose
focus from the JTextField, which may be counter-intuitive.  I may have to
revisit this later.