+ Feature: Added "Equipment" context menu in lounge, for burst MGs, hot loading and searchlights (Covers SF features #1399 and #1418)
Now, when units are right-clicked on, if relevant options are enabled, an
"Equipment" menu item is added and it will have submenus for setting the burst
state on MGs, hot loading on LRMs, and searchlights, if relevant.

This essentially adds a way to affect the Equipment tab in the custom unit
dialog, since editing equipment in that dialog is hairy for multiple units.