+ Bug: UnitEditorDialog always sets critical hit on ASF life support
ASF.setLifeSupport takes a boolean that is true if the unit has life support
and false if the units life support is broken.  This differs from similar
methods in ASF (like setSensorHits, setGearHits, setEngineHits, etc.) and also
differs from how the life support hits work for Mechs.  Because of this, there
was a logic bug where it was setting it to true when the hits were > 0 instead
instead of when they are equal to 0.