+ Bug: Clicking Walk from Jump, would still plot jumping movement
Since clear() (and pressing the escape key) no longer resets the movement mode,
there was a bug introduced where clicking the Walk button would change it
back to walk mode.  The problem is, when walk is pressed, if the mode was
jump or swim, clear is called, then the gear is set to land (for walk).
The problem was, when clear() was called, since the gear jump jump,
clear() added a START_JUMP movestep, which made the path considered jumping.