+ Feature: Non-Infantry Weapon Damage Against Infantry changes now reported
I found the reports for damage against infantry in a building to be fairly
confusing, and the main reason is that the weapon damage ends up being
vastly different against infantry (For instance, firing a Clan Large Laser
at infantry in a building, once might expect the infantry to still take
decent damage, but since it's a direct fire weapon the 10 damage is reduced
to 1, so if the building absorbs 50% or more damage, then none is dealt).

Now, this damage adjustment is reported, making damage against infantry much
clearer.