Issue with moving into hexes with hidden units
I had written some code to report when a unit tries to move into a hex
with a hidden unit and it creates a stacking violation.  Unfortunately, this
code was never firing.  The client sends a move path with the last forward
step, but then the server clips it off, because it knows that it's impossible
(the client didn't know about hte hidden unit, so didn't know about the
stacking violation).

I have fixed this by not clipping off impossible steps for the movepath.  This
is ok, because illegal steps are not processed anyways (however, I did have to
move some code in MoveStep).  So now, that reporting code properly fires.  It
should now be much clearer just what happened when a unit tries to move into
a hex with a hidden unit, and the hidden unit doesn't take the point-blank
shot.