Added ImageUtil.getScaledImage and updated calls.
Instead of using Image.getScaledInstance, which is apparently outdated and slower,
I have used Graphics2D.drawImage, with bicubic interpolation.  This keeps more
details when scaling down, and also runs faster.  I've tried to convert most calls
to ImageUtil.getScaledImage, so if we need to change the method, it can be done
easier.