Found some sprites using old code for transparency
At some point, we had some issues creating transparent images for sprites.
A method was developed where an image was created, then filled with a defined
TRANSLUCENT hex, the desired image was draw on, and then MM's KeyAlphaFilter
was applied.  This was found to cause issues with transparency in PNGs.  If
a unit had supplied transparency, it would get a pink halo.  This old method
is no longer required and there's a much simpler way now.