Added MegaMekFile, a wraper around java.io.File
Essentially, it is used to determine if the specified file exists in the
useredata directory, and if it does use that file instead.  It's designed
to replace most usages of java.io.File.

Unfortunately, because java requires the super constructor to get the first
line, I can't really extend File and instead have to wrap it.