Fixed spurs on flight path bends
This is somewhat difficult for me to describe, but what is happening is that
we need to have a closed geometric shape in order to get the fill and stroke.
We can't just figure out the hex center points along the path and draw a line.

So, we draw the path along the hex centers, but then use trig to get the
"outside" and "inside" points of the "path" shape.  First you trace down the
outside, then trace up the inside.

For bends, we use the angle between two hexes.  The problem is, when going
around bends, because we have hexes that are 60 degrees to a facing, we won't
get two angles that line up, so we have to "smooth out" corners a bit. I've
added code for doing this.