Bug: Climb Mode toggle appears to have three modes
If you click Climb Mode, it cycles through three options: nothing display,
"Go Thru", and "Climb Up".  The reason there are three is because, if you
click on the climb mode button, and the last step is a climb mode step, it
just removes that step.  Instead, I changed it so that it replaces that last
step with the step of the other  type, so once you press climb mode, it cycles
through the two actual modes.  I think this is less confusing to the user.