Made climb mode an option, default to true
Now, the default climb mode is settable via an advanced option and because of
this, it now resets to the default value at each round.  The default value has
been changed from "Go Thru" to "Climb Up".   This is because "Climb Up" is
generally the option you want to use for bridges.