Client Settings Advanced  display names & tooltips
I added support for reading i18n properties files for names and tooltips for
the entries in the list in the Client Settings > Advanced tab.  I did my best
to add useful names and tooltips, but many of the names still aren't the most
useful, and the number of tooltips is small.