Fixes #399: Iso bridges not darkened by night
There was code in drawHex that would go through the hex image and adjust each
pixel's color based on the light level, however this code never fired for
the orthographic images drawn in isometric mode.  I pulled the code out into
a method and now it's also applied to the bridge image in drawOrthographic.

Unfortunately, the code requires a BufferedImage and the orthogrpahic images
aren't guaranteed to be ARGB BufferedImages so I added a call to
ImageUtil.createAcceleratedImage.  This isn't ideal, but there shouldn't be
many orthographic images right now, so it works.