Added CreateImageAtlases app
This program will scan the unit and hexes image directories and store any
images that are the proper size (84x72) into an atlas.  It will also save a
file that maps the filenames of images stored to their atlas and place within
the atlas, as well as a list of files that have been stored in an atlas.
The former can be used by an ImageLoader within MM to make use of the atlas
and the later can be used to delete files that are no longer necessary for a
release.