Force font colors in colored BV tooltip tables
There are several HTML tables that get defined for board view tooltips that use
the TABLE BGCOLOR attribute.  These will set the background color for the table
and it works fine if the font text remains black.  There are instances however
where the user's configuration can change the colors for the tooltips (I assume
these are settigns inherited from the operating systems appearance settings).
If the tooltip text gets set to be something other than black, the text in the
table may not be readable.  To fix this, in any case where we force a
background color, I also force a font color of black.