Added ability for BoardEditor to load invalid maps
It becomes a pain to fix invalid maps if the board editor can't load them.
Now, invalid maps can be loaded, but a message is displayed that the map
is invalid.

Really, we should have a lot more support for indicating why the map is invalid
but I'm going to save that for a larger enhancement to the board editor.