Added Targetable.isEnemyOf(Entity)
There are situations where an attacking Entity wants to know if a Targetable
should be considered an enemy, to see if the attack is valid.  Entity already
implements this function and I suppose it wasn't added to Targetable because
Entity's version uses IPlayer to make comparison, and targets like buildings
and hexes don't have owning Players.  However, in this case, we can just return
true.  By adding Targetable.isEnemyOf(Entity), it allows us to avoid some
awkward checks for Targetable type.