Fix #487 Added FixCRLF to nix-bundle task
Apparently, the copy task can introduce windows line-endings in files.  This
doesn't appear to be documented from what I can see, but I think the effects
in practice.  This causes problems when releases are packaged on windows,
as the UNIX startup.sh ends up failing because the carriage return on the
 #!bin/bash^M line causes the shell definition to fail.

To fix this, I used the ant FixCRLF task, which allows one to set the line
endings.