Account for movement by tracked non-QuadVees with destroyed gyros.
Legal moves for 'Mechs with destroyed gyros, as the code now stands:
1. Prone 'Mechs can only change one hex side and cannot stand.
2. Non-prone 'Mechs that do not have tracks or wheels (QuadVee or
otherwise) cannot expend MP even to change hex side (this situation
should be impossible to achieve in-game, but you can get into it with MM
if you edit the damage to destroy the gyro without setting it to deploy
prone).
3. Non-prone, non-QuadVee 'Mechs with tracks can use tracks to move
normally.
4. Non-prone QuadVees in vehicle mode or converting to 'Mech mode move
normally.
5. Non-prone QuadVees in 'Mech mode can spend MP to convert to vehicle
mode, but that is the end of the movement turn.