Added BV listing to the on map unit tooltip.
Adds a feature requested in #479. Allows a user to quickly see the current battle value of a unit.  The listing shows the current bv, the initial bv, and the percentage of bv remaining. If playing double blind and the "Hide Battle Value reports" option is selected, the bv listings will not be visible on enemy units, but will be visible on your units.

The `Entity` class didn't have any way to keep track of a units initial BV, so I had to add it. I added a getter and a setter for `initialBV` in the `Entity` class, and then added a call to `Entity.setInitialBV(int)` in `Game.addEntity(Entity, boolean)`.