Support multiple resolutions for splash image
Now, the third and fourth background_image tag for the main menu border will
be used to find a UHD and <=HD version of the splash image.  That is, if the
largest monitor resolution is <1600 width, the <=HD splash image is used, if
the largest monitor resolution is >1920 in width, the UHD splash is used.
This allows us to set image sizes that scale with monitor resolution.