Adjusted button widths for main menu
This should keep buttons a bit wider, and also prevent them from being too
narrow.  The updates set min and preferred dimensions.  Two different dims
are computed: 30% of the width of the splash image, and also based on the
width of the text.  The widest dim is picked as the min and preferred.