Fail to load images that don't exist
Previously, TileMapImageLoader would still return a non-null but empty Image
if the file it was trying to load didn't exist.  I think it was assumed that
Toolkit.getImage(File) would return null if the file doesn't exist, but that
isn't the case.  Instead it returns a blank image, which means that the loading
appears to work.

I added an explicit check for a non-existant file, and a null is returned if
the file doesn't exist. This will help diagnose errors in tilesets, as at
least a message should get logged.