Fighter squadron bombing attacks, take 2.
VTOL bombing required applying bomb choices separately from setting up
fighters during the exchange phase, and the order is important; bomb
choices must be applied before setting up squadron weapon groups.

There was also a problem with squadrons not being able to find the ammo
for external weapons because it was stored as one-shot ammo to keep from
taking up a critical spot in LAM bomb bays which was resolved by storing
it in the same location as the weapon and checking whether it is bomb
ammo before assigning a critical slot in a mech.