Account for variable weights of infantry transport bays in InfantryBay based on motive type.
Previously it treated all infantry platoons the same, but different
types have different tonnage requirements to account for maintenance
facilities and sleeping quarters. Requiring a specific infantry bay type
for the infantry is rather burdonsome, as it would require creating a
custom unit just to carry a different type of infantry. Instead I
followed what I consider a common sense approach and assume the bay has
a factory setting (defaulting to foot) but can be reconfigured on the
fly to accomodate other types. Infantry is already fairly abstracted in
the rules.